requirements.txt
fastapi[all]~=0.104.0
redis~=5.0.1
uvicorn
pydantic
httpx~=0.25.0
python-dotenv~=1.0.0